redefine munmap parameter type from int to uint64

+++ user/user.h
@@ -24,7 +24,7 @@ char* sbrk(int);
 int sleep(int);
 int uptime(void);
 char *mmap(void *, uint64, int, int, int, uint64);
-int munmap(void *, int);
+int munmap(void *, uint64);
 
 // ulib.c
 int stat(const char*, struct stat*);
